requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
amqp==5.0.6; python_version >= '3.6'
asgiref==3.4.1; python_version >= '3.6'
billiard==3.6.4.0
celery==5.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-cors-headers==3.7.0
django==3.2.6
djangorestframework==3.12.4
kombu==5.1.0; python_version >= '3.6'
# pillow==8.3.1
prompt-toolkit==3.0.19; python_full_version >= '3.6.1'
# psycopg2-binary==2.9.1
pytz==2021.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.5
python-dotenv
